extract RemoveExcludedCheckersCompilerPass
